requirements.txt
httpx
pyserial
pyyaml
uvicorn[standard]
fastapi
websockets
numpy
opencv-python-headless
